+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 3. Fix Katherine and Fiora attributes (life and attack)
 4. Better implementation of levitation orb
 5. Better implementation of weapon
+6. Use data generator for all data(pack) generation
+7. Use loot table for advancement rewards, no longer use mcfunction
+8. Fix the disappearing problem when inserting medieval music disc.
+9. Fix the mineable problem of amat, boss, and the hardened metal blocks.